scripts/cloudfunctions/recommender-checker/requirements.txt
google-cloud-asset==3.7.1
google-cloud-recommender==2.5.1
